+++ database/migrations/2025_10_20_000008_create_user_show_statuses_table.php
@@ -11,13 +11,15 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('user_show_statuses', function (Blueprint $table) {
-            $table->id()->index();
-            $table->foreignId('show_id')->constrained('shows')->onDelete('cascade');
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-            $table->string('status');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('user_show_statuses')) {
+            Schema::create('user_show_statuses', function (Blueprint $table) {
+                $table->id()->index();
+                $table->foreignId('show_id')->constrained('shows')->onDelete('cascade');
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+                $table->string('status');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
